move concerns to the model/concerns
