+++ app/models/concerns/trashable.rb
